requirements.txt
pandas
pillow
streamlit
google-generativeai
flask